pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot==13.7",
    "redis>=5.2.1",
    "sendgrid>=6.11.0",
    "telegram>=0.0.1",
    "twilio>=9.4.4",
    "yt-dlp>=2025.1.26",
]
